+++ packages/plugin-hardhat/hardhat.config.js
@@ -4,10 +4,16 @@ module.exports = {
   solidity: {
     compilers: [
       {
-        version: '0.5.15',
+        version: '0.8.9',
       },
       {
-        version: '0.6.7',
+        version: '0.7.6',
+      },
+      {
+        version: '0.6.12',
+      },
+      {
+        version: '0.5.17',
       },
     ],
   },