+++ packages/plugin-hardhat/src/utils/deploy-impl.ts
@@ -3,7 +3,6 @@ import type { ContractFactory } from 'ethers';
 
 import {
   Manifest,
-  ValidationOptions,
   assertUpgradeSafe,
   assertStorageUpgradeSafe,
   fetchOrDeploy,
@@ -13,11 +12,12 @@ import {
   getUnlinkedBytecode,
   getVersion,
   inferProxyKind,
-  withValidationDefaults,
   setProxyKind,
+  ValidationOptions,
 } from '@openzeppelin/upgrades-core';
 
 import { deploy } from './deploy';
+import { Options, withDefaults } from './options';
 import { readValidations } from './validations';
 
 interface DeployedImpl {
@@ -28,13 +28,14 @@ interface DeployedImpl {
 export async function deployImpl(
   hre: HardhatRuntimeEnvironment,
   ImplFactory: ContractFactory,
-  opts: ValidationOptions,
+  opts: Options,
   proxyAddress?: string,
 ): Promise<DeployedImpl> {
   const { provider } = hre.network;
   const validations = await readValidations(hre);
   const unlinkedBytecode = getUnlinkedBytecode(validations, ImplFactory.bytecode);
-  const version = getVersion(unlinkedBytecode, ImplFactory.bytecode);
+  const encodedArgs = ImplFactory.interface.encodeDeploy(opts.constructorArgs);
+  const version = getVersion(unlinkedBytecode, ImplFactory.bytecode, encodedArgs);
   const layout = getStorageLayout(validations, version);
 
   if (opts.kind === undefined) {
@@ -45,7 +46,7 @@ export async function deployImpl(
     await setProxyKind(provider, proxyAddress, opts);
   }
 
-  const fullOpts = withValidationDefaults(opts);
+  const fullOpts = withDefaults(opts);
 
   assertUpgradeSafe(validations, version, fullOpts);
 
@@ -57,7 +58,7 @@ export async function deployImpl(
   }
 
   const impl = await fetchOrDeploy(version, provider, async () => {
-    const deployment = await deploy(ImplFactory);
+    const deployment = await deploy(ImplFactory, ...fullOpts.constructorArgs);
     return { ...deployment, layout };
   });