+++ packages/plugin-truffle/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.10.0
 
 - Infer whether a proxy should be UUPS or Transparent. ([#441](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/441))
+- Add implementation constructor arguments to initialize immutable variables. ([#433](https://github.com/OpenZeppelin/openzeppelin-upgrades/pull/433))
 
 ## 1.9.1 (2021-08-31)